# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Orion — Declarative Services Runtime API",
    "description": "Declarative services runtime platform.\n\n**Authentication.** The admin API (`/api/v1/admin/*`), the Prometheus endpoint (`/metrics`), and the trace read endpoints (`/api/v1/admin/traces*`) require an admin API key when `admin_auth.enabled` is true — the default in the shipped Helm chart and HA compose files. Operations that need it carry a `security` block; see the `admin_bearer` / `admin_api_key` schemes for how the key is presented and for the `admin_auth.header` config key that selects between them.\n\nThe data plane (`/api/v1/data/{channel}`) and the health probes are deliberately unauthenticated: channel-level access control is expressed through each channel's `validation_logic` and CORS configuration.\n\n**Errors.** Every non-2xx response uses the `ErrorResponse` envelope: `{\"error\": {\"code\", \"message\", \"request_id\"}}`, plus `details` on validation failures.",
    "license": {
      "name": "Apache-2.0"
    },
    "version": "1.0.0"
  },
  "paths": {
    "/api/v1/admin/audit-logs": {
      "get": {
        "tags": [
          "Audit"
        ],
        "operationId": "list_audit_logs",
        "parameters": [
          {
            "name": "offset",
            "in": "query",
            "description": "Pagination offset (default 0)",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Page size, clamped to [1, 1000] (default 50)",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "action",
            "in": "query",
            "description": "Exact-match filter on the action (e.g. `create`, `activate`, `delete`)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "resource_type",
            "in": "query",
            "description": "Exact-match filter on the resource type (`workflow`, `channel`, `connector`, `engine`)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "resource_id",
            "in": "query",
            "description": "Exact-match filter on the resource ID",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "principal",
            "in": "query",
            "description": "Exact-match filter on the acting principal",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "start_time",
            "in": "query",
            "description": "Inclusive lower bound on `created_at`, RFC 3339",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "end_time",
            "in": "query",
            "description": "Exclusive upper bound on `created_at`, RFC 3339",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated audit log entries",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedEnvelope_AuditLogEntryResponse"
                }
              }
            }
          },
          "400": {
            "description": "Unknown query parameter or malformed timestamp",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid admin API key. Only returned when `admin_auth.enabled` is true.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected internal error (`INTERNAL_ERROR`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "admin_bearer": []
          },
          {
            "admin_api_key": []
          }
        ]
      }
    },
    "/api/v1/admin/backups": {
      "get": {
        "tags": [
          "Backups"
        ],
        "operationId": "list_backups",
        "responses": {
          "200": {
            "description": "List of backup files in the configured backup directory",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataEnvelope_Vec_BackupListItem"
                }
              }
            }
          },
          "400": {
            "description": "Unavailable in cluster mode — backups are node-local files",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid admin API key. Only returned when `admin_auth.enabled` is true.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected internal error (`INTERNAL_ERROR`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "admin_bearer": []
          },
          {
            "admin_api_key": []
          }
        ]
      },
      "post": {
        "tags": [
          "Backups"
        ],
        "operationId": "create_backup",
        "responses": {
          "200": {
            "description": "Backup created (SQLite only — VACUUM INTO a timestamped file)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataEnvelope_BackupFile"
                }
              }
            }
          },
          "400": {
            "description": "Backup unavailable (non-SQLite backend, or cluster mode — use managed-DB snapshots/PITR)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid admin API key. Only returned when `admin_auth.enabled` is true.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected internal error (`INTERNAL_ERROR`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "admin_bearer": []
          },
          {
            "admin_api_key": []
          }
        ]
      }
    },
    "/api/v1/admin/channels": {
      "get": {
        "tags": [
          "Channels"
        ],
        "operationId": "list_channels",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "channel_type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "protocol",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "sort_by",
            "in": "query",
            "description": "Column to sort by: priority (default), name, status, channel_type, protocol, created_at, updated_at.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort_order",
            "in": "query",
            "description": "Sort direction: asc or desc (default).",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of channels",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedEnvelope_ChannelResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid admin API key. Only returned when `admin_auth.enabled` is true.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected internal error (`INTERNAL_ERROR`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "admin_bearer": []
          },
          {
            "admin_api_key": []
          }
        ]
      },
      "post": {
        "tags": [
          "Channels"
        ],
        "operationId": "create_channel",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateChannelRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Channel created as draft",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataEnvelope_ChannelResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid admin API key. Only returned when `admin_auth.enabled` is true.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Channel id already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected internal error (`INTERNAL_ERROR`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "admin_bearer": []
          },
          {
            "admin_api_key": []
          }
        ]
      }
    },
    "/api/v1/admin/channels/import": {
      "post": {
        "tags": [
          "Channels"
        ],
        "operationId": "import_channels",
        "parameters": [
          {
            "name": "dry_run",
            "in": "query",
            "description": "When true, validate each item and report what would happen without\nwriting. Probes for conflicts against stored rows and for duplicates\nwithin the batch (R15).",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/CreateChannelRequest"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Import results with counts (or would-be results when ?dry_run=true). Each item is handled independently: a malformed or conflicting item becomes one entry in `errors` and the rest of the batch still applies. Dry-run additionally probes for id conflicts against stored rows and duplicates within the batch, without writing.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataEnvelope_ImportResult"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid admin API key. Only returned when `admin_auth.enabled` is true.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected internal error (`INTERNAL_ERROR`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "admin_bearer": []
          },
          {
            "admin_api_key": []
          }
        ]
      }
    },
...
    "/api/v1/admin/connectors/circuit-breakers": {
      "get": {
        "tags": [
          "Connectors"
        ],
        "operationId": "list_circuit_breakers",
        "responses": {
          "200": {
            "description": "Circuit breaker states",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataEnvelope_CircuitBreakerStates"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid admin API key. Only returned when `admin_auth.enabled` is true.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected internal error (`INTERNAL_ERROR`)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "admin_bearer": []
          },
          {
            "admin_api_key": []
          }
        ]
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "admin_api_key": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Alternative to `admin_bearer`, active when `admin_auth.header` names a header other than `Authorization`. The raw key is sent as that header's value with no `Bearer ` prefix. `X-API-Key` is the conventional choice and is what this document shows, but the header name is deployment-specific — substitute whatever `admin_auth.header` is set to. Exactly one of the two schemes is live at a time; they are listed as alternatives because this document is generated without knowledge of a deployment's config."
      },
      "admin_bearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Admin API key presented as `Authorization: Bearer <key>` — the default. Enforced on `/api/v1/admin/*`, `/metrics`, and `/api/v1/admin/traces*` whenever `admin_auth.enabled` is true, which the shipped Helm chart and HA compose files set. Keys come from `admin_auth.api_keys`, either in plaintext or as `sha256:<64-hex>` digests. The data plane (`POST /api/v1/data/{channel}`) is not covered by this scheme."
      }
    }
  },
...
}